requirements.txt
albumentations==1.4.4
easyocr==1.7.1
loguru==0.7.2
numpy==1.26.4
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
Pillow==10.3.0
pycocotools==2.0.7
pytesseract==0.3.10
python-dotenv==1.0.1
PyYAML==6.0.1
tesserocr==2.7.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
